refactor(events): Rename private module engbert to microsaccades
