+++ src/pymovements/events/__init__.py
@@ -60,7 +60,7 @@
     pymovements.events.event_properties.amplitude
     pymovements.events.event_properties.peak_velocity
 """
-from pymovements.events.detection.engbert import microsaccades
+from pymovements.events.detection.microsaccades import microsaccades
 from pymovements.events.detection.idt import idt
 from pymovements.events.detection.ivt import ivt
 from pymovements.events.event_processing import EventGazeProcessor
